Extract a common base class for all FsHelper classes based on the hadoop filesystem.  Added a fix for the hadoop issue: https://issues.apache.org/jira/browse/HADOOP-12169
